+++ .ci.sh
@@ -1,6 +1,4 @@
-shCiPre() {(set -e
-# this function will run pre-ci
+shCiPreCustom() {(set -e
+# this function will run pre-ci-custom
     shCiPreMyci
 )}
-
-. ./myci2.sh